// roppi-frontend/src/types/producto/genericoAtributos.types.ts
export interface GenericoXColor { 
    id_color: number; }
export interface GenericoXMaterial { 
    id_material: number; costo_extra: number; }
export interface GenericoXTamano { 
    id_tamano: number; alto: number; ancho: number; }
export interface GenericoXPersonalizacion { 
    id_tipo_personalizacion: number; costo_extra: number; }